backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8
# retellai package removed - using httpx for direct API calls instead
